Cleanup Deprecrated File APIs
