+++ nodestream/pipeline/extractors/__init__.py
@@ -1,14 +1,13 @@
 from .apis import SimpleApiExtractor
 from .extractor import Extractor
-from .files import FileExtractor, RemoteFileExtractor
+from .files import FileExtractor
 from .iterable import IterableExtractor
 from .ttls import TimeToLiveConfigurationExtractor
 
 __all__ = (
     "Extractor",
     "IterableExtractor",
-    "FileExtractor",
-    "RemoteFileExtractor",
     "TimeToLiveConfigurationExtractor",
     "SimpleApiExtractor",
+    "FileExtractor",
 )
