+++ nodestream/pipeline/extractors/stores/aws/__init__.py
@@ -1,5 +1,4 @@
 from .athena_extractor import AthenaExtractor
 from .dynamodb_extractor import DynamoDBExtractor
-from .s3_extractor import S3Extractor
 
-__all__ = ("AthenaExtractor", "S3Extractor", "DynamoDBExtractor")
+__all__ = ("AthenaExtractor", "DynamoDBExtractor")
