+++ playground_project/test.yaml
@@ -1,4 +1,5 @@
 - implementation: nodestream.pipeline.extractors:FileExtractor
+  factory: local
   arguments:
     globs:
       - people/*.json
