+++ tests/integration/fixtures/pipelines/source_match_only.yaml
@@ -1,4 +1,5 @@
 - implementation: nodestream.pipeline.extractors:FileExtractor
+  factory: local
   arguments:
     globs:
       - tests/data/source_match_only_data.csv
