+++ tests/unit/pipeline/extractors/test_files.py
@@ -8,15 +8,22 @@ from tempfile import NamedTemporaryFile, TemporaryDirectory
 import pandas as pd
 import pytest
 import yaml
-from hamcrest import assert_that, contains_string, equal_to, has_length
+from hamcrest import (
+    assert_that,
+    contains_string,
+    equal_to,
+    has_items,
+    has_key,
+    has_length,
+    not_,
+)
+from moto import mock_aws
 
 from nodestream.pipeline.extractors.files import (
     FileExtractor,
     FileSource,
     LocalFileSource,
-    RemoteFileExtractor,
     RemoteFileSource,
-    UnifiedFileExtractor,
 )
 
 SIMPLE_RECORD = {"record": "value"}
@@ -148,7 +155,7 @@ async def test_unsupported_file(unsupported_file):
 
 @pytest.mark.asyncio
 async def test_supported_file_unified(unsupported_file):
-    subject = UnifiedFileExtractor.from_file_data(
+    subject = FileExtractor.from_file_data(
         [{"type": "local", "globs": [str(unsupported_file)]}]
     )
     results = [r async for r in subject.extract_records()]
@@ -164,7 +171,7 @@ async def test_json_formatting(json_file):
 
 @pytest.mark.asyncio
 async def test_json_formatting_unified(json_file):
-    subject = UnifiedFileExtractor.from_file_data(
+    subject = FileExtractor.from_file_data(
         [{"type": "local", "globs": [str(json_file)]}]
     )
     results = [r async for r in subject.extract_records()]
@@ -180,7 +187,7 @@ async def test_csv_formatting(csv_file):
 
 @pytest.mark.asyncio
 async def test_csv_formatting_unified(csv_file):
-    subject = UnifiedFileExtractor.from_file_data(
+    subject = FileExtractor.from_file_data(
         [{"type": "local", "globs": [str(csv_file)]}]
     )
     results = [r async for r in subject.extract_records()]
@@ -196,7 +203,7 @@ async def test_txt_formatting(txt_file):
 
 @pytest.mark.asyncio
 async def test_txt_formatting_unified(txt_file):
-    subject = UnifiedFileExtractor.from_file_data(
+    subject = FileExtractor.from_file_data(
         [{"type": "local", "globs": [str(txt_file)]}]
     )
     results = [r async for r in subject.extract_records()]
@@ -212,7 +219,7 @@ async def test_jsonl_formatting(jsonl_file):
 
 @pytest.mark.asyncio
 async def test_jsonl_formatting_unified(jsonl_file):
-    subject = UnifiedFileExtractor.from_file_data(
+    subject = FileExtractor.from_file_data(
         [{"type": "local", "globs": [str(jsonl_file)]}]
     )
     results = [r async for r in subject.extract_records()]
@@ -228,7 +235,7 @@ async def test_yaml_formatting(yaml_file):
 
 @pytest.mark.asyncio
 async def test_yaml_formatting_unified(yaml_file):
-    subject = UnifiedFileExtractor.from_file_data(
+    subject = FileExtractor.from_file_data(
         [{"type": "local", "globs": [str(yaml_file)]}]
     )
     results = [r async for r in subject.extract_records()]
@@ -244,7 +251,7 @@ async def test_parquet_formatting(parquet_file):
 
 @pytest.mark.asyncio
 async def test_parquet_formatting_unified(parquet_file):
-    subject = UnifiedFileExtractor.from_file_data(
+    subject = FileExtractor.from_file_data(
         [{"type": "local", "globs": [str(parquet_file)]}]
     )
     results = [r async for r in subject.extract_records()]
@@ -260,7 +267,7 @@ async def test_gzip_formatting(gzip_file):
 
 @pytest.mark.asyncio
 async def test_gzip_formatting_unified(gzip_file):
-    subject = UnifiedFileExtractor.from_file_data(
+    subject = FileExtractor.from_file_data(
         [{"type": "local", "globs": [str(gzip_file)]}]
     )
     results = [r async for r in subject.extract_records()]
@@ -276,7 +283,7 @@ async def test_bz2_formatting(bz2_file):
 
 @pytest.mark.asyncio
 async def test_bz2_formatting_unified(bz2_file):
-    subject = UnifiedFileExtractor.from_file_data(
+    subject = FileExtractor.from_file_data(
         [{"type": "local", "globs": [str(bz2_file)]}]
     )
     results = [r async for r in subject.extract_records()]
@@ -286,7 +293,7 @@ async def test_bz2_formatting_unified(bz2_file):
 def test_declarative_init(
     fixture_directory, csv_file, json_file, txt_file, jsonl_file, yaml_file
 ):
-    subject = FileExtractor.from_file_data(globs=[f"{fixture_directory}/**"])
+    subject = FileExtractor.local(globs=[f"{fixture_directory}/**"])
     assert_that(list(subject.file_sources[0].paths), has_length(5))
 
 
@@ -299,14 +306,14 @@ async def test_remote_file_extractor_extract_records(mocker, httpx_mock):
             method="GET",
             json=SIMPLE_RECORD,
         )
-    subject = RemoteFileExtractor.from_file_data(urls=files)
+    subject = FileExtractor.remote(urls=files)
     results = [r async for r in subject.extract_records()]
     assert_that(results, equal_to([SIMPLE_RECORD, SIMPLE_RECORD]))
 
 
 @pytest.mark.asyncio
 async def test_no_files_found_from_local_source(mocker):
-    subject = UnifiedFileExtractor([LocalFileSource([])])
+    subject = FileExtractor([LocalFileSource([])])
     subject.logger = mocker.Mock()
     results = [r async for r in subject.extract_records()]
     assert_that(results, equal_to([]))
@@ -317,7 +324,7 @@ async def test_no_files_found_from_local_source(mocker):
 
 @pytest.mark.asyncio
 async def test_no_files_found_from_remote_source(mocker):
-    subject = UnifiedFileExtractor([RemoteFileSource([], 10)])
+    subject = FileExtractor([RemoteFileSource([], 10)])
     subject.logger = mocker.Mock()
     results = [r async for r in subject.extract_records()]
     assert_that(results, equal_to([]))
@@ -372,7 +379,7 @@ async def test_unified_extractor_multiple_source_types(json_file, csv_file, http
             json=SIMPLE_RECORD,
         )
 
-    subject = UnifiedFileExtractor.from_file_data(
+    subject = FileExtractor.from_file_data(
         [
             {"type": "local", "globs": [str(json_file)]},
             {"type": "local", "globs": [str(csv_file)]},
@@ -381,3 +388,290 @@ async def test_unified_extractor_multiple_source_types(json_file, csv_file, http
     )
     results = [r async for r in subject.extract_records()]
     assert_that(results, equal_to([SIMPLE_RECORD] * 4))
+
+
+BUCKET_NAME = "bucket"
+PREFIX = "prefix"
+NUM_OBJECTS = 10
+
+
+@pytest.fixture
+def s3_client():
+    with mock_aws():
+        import boto3
+
+        yield boto3.client("s3", region_name="us-east-1")
+
+
+@pytest.fixture
+def subject(mocker):
+    from nodestream.pipeline.extractors.credential_utils import AwsClientFactory
+
+    mocker.patch.object(AwsClientFactory, "make_client")
+    return FileExtractor.s3(bucket="bucket", prefix=PREFIX)
+
+
+@pytest.fixture
+def subject_with_populated_objects(subject, s3_client):
+    subject.file_sources[0].s3_client = s3_client
+    s3_client.create_bucket(Bucket=BUCKET_NAME)
+    s3_client.put_object(Bucket=BUCKET_NAME, Key="notprefixed", Body="hello".encode())
+    for i in range(NUM_OBJECTS):
+        s3_client.put_object(
+            Bucket=BUCKET_NAME,
+            Key=f"{PREFIX}/foo/{i}.json",
+            Body=json.dumps({"hello": i}),
+        )
+    return subject
+
+
+@pytest.fixture
+def subject_with_populated_objects_and_no_extension(subject, s3_client):
+    subject.file_sources[0].s3_client = s3_client
+    s3_client.create_bucket(Bucket=BUCKET_NAME)
+    s3_client.put_object(Bucket=BUCKET_NAME, Key="notprefixed", Body="hello".encode())
+    for i in range(NUM_OBJECTS):
+        s3_client.put_object(
+            Bucket=BUCKET_NAME,
+            Key=f"{PREFIX}/foo/{i}",
+            Body=json.dumps({"hello": i}),
+        )
+    return subject
+
+
+@pytest.fixture
+def subject_with_populated_csv_objects(subject, s3_client):
+    subject.file_sources[0].s3_client = s3_client
+    s3_client.create_bucket(Bucket=BUCKET_NAME)
+    s3_client.put_object(Bucket=BUCKET_NAME, Key="notprefixed", Body="hello".encode())
+    for i in range(NUM_OBJECTS):
+        s3_client.put_object(
+            Bucket=BUCKET_NAME,
+            Key=f"{PREFIX}/bar/{i}.csv",
+            Body=f"column1,column2\nvalue{i},value{i+10}",
+        )
+    return subject
+
+
+@pytest.fixture
+def subject_with_populated_jsonl_objects(subject, s3_client):
+    subject.file_sources[0].s3_client = s3_client
+    s3_client.create_bucket(Bucket=BUCKET_NAME)
+    for i in range(1):
+        s3_client.put_object(
+            Bucket=BUCKET_NAME,
+            Key=f"{PREFIX}/bar/{i}.jsonl",
+            Body='{"test":"test"}\n{"test2":"test2"}',
+        )
+    return subject
+
+
+@pytest.fixture
+def subject_with_populated_text_objects(subject, s3_client):
+    subject.file_sources[0].s3_client = s3_client
+    s3_client.create_bucket(Bucket=BUCKET_NAME)
+    for i in range(1):
+        s3_client.put_object(
+            Bucket=BUCKET_NAME,
+            Key=f"{PREFIX}/bar/{i}.txt",
+            Body="test\ntest2\n",
+        )
+    return subject
+
+
+@pytest.fixture
+def subject_with_archieved_objects(subject_with_archiving_enabled, s3_client):
+    for i in range(NUM_OBJECTS):
+        s3_client.put_object(
+            Bucket=BUCKET_NAME,
+            Key=f"archieve/{i}.json",
+            Body=json.dumps({"hello": i}),
+        )
+    return subject_with_archiving_enabled
+
+
+@pytest.fixture
+def subject_with_populated_and_gz_compressed_objects(subject, s3_client):
+    subject.file_sources[0].s3_client = s3_client
+    s3_client.create_bucket(Bucket=BUCKET_NAME)
+
+    for i in range(NUM_OBJECTS):
+        body_data = json.dumps({"hello": i}).encode()
+        gzipped_body = gzip.compress(body_data)
+
+        s3_client.put_object(
+            Bucket=BUCKET_NAME,
+            Key=f"{PREFIX}/foo/{i}.json.gz",
+            Body=gzipped_body,
+        )
+
+    return subject
+
+
+@pytest.fixture
+def subject_with_populated_and_bz2_compressed_objects(subject, s3_client):
+    subject.file_sources[0].s3_client = s3_client
+    s3_client.create_bucket(Bucket=BUCKET_NAME)
+
+    for i in range(NUM_OBJECTS):
+        body_data = json.dumps({"hello": i}).encode()
+        bz2_body = bz2.compress(body_data)
+
+        s3_client.put_object(
+            Bucket=BUCKET_NAME,
+            Key=f"{PREFIX}/foo/{i}.json.bz2",
+            Body=bz2_body,
+        )
+
+    return subject
+
+
+@pytest.fixture
+def subject_with_populated_double_compressed_objects(subject, s3_client):
+    subject.file_sources[0].s3_client = s3_client
+    s3_client.create_bucket(Bucket=BUCKET_NAME)
+
+    for i in range(NUM_OBJECTS):
+        body_data = json.dumps({"hello": i}).encode()
+        gzipped_body = gzip.compress(body_data)
+        double_zipped_body = bz2.compress(gzipped_body)
+
+        s3_client.put_object(
+            Bucket=BUCKET_NAME,
+            Key=f"{PREFIX}/foo/{i}.json.gz.bz2",
+            Body=double_zipped_body,
+        )
+
+    return subject
+
+
+@pytest.mark.asyncio
+async def test_s3_extractor_properly_loads_csv_files(
+    subject_with_populated_csv_objects,
+):
+    expected_results = [
+        {"column1": f"value{i}", "column2": f"value{i+10}"} for i in range(NUM_OBJECTS)
+    ]
+    results = [
+        result async for result in subject_with_populated_csv_objects.extract_records()
+    ]
+    assert_that(results, has_length(NUM_OBJECTS))
+    assert_that(results, has_items(*expected_results))
+
+
+@pytest.mark.asyncio
+async def test_s3_extractor_properly_loads_jsonl_files(
+    subject_with_populated_jsonl_objects,
+):
+    expected_results = [{"test": "test"}, {"test2": "test2"}]
+    results = [
+        result
+        async for result in subject_with_populated_jsonl_objects.extract_records()
+    ]
+    assert_that(results, has_length(2))
+    assert_that(results, has_items(*expected_results))
+
+
+@pytest.mark.asyncio
+async def test_s3_extractor_properly_loads_text_files(
+    subject_with_populated_text_objects,
+):
+    expected_results = [{"line": "test"}, {"line": "test2"}]
+    results = [
+        result async for result in subject_with_populated_text_objects.extract_records()
+    ]
+    assert_that(results, has_length(2))
+    assert_that(results, has_items(*expected_results))
+
+
+@pytest.fixture
+def subject_with_archiving_enabled(subject_with_populated_objects):
+    subject_with_populated_objects.file_sources[0].archive_dir = "archive"
+    return subject_with_populated_objects
+
+
+@pytest.mark.asyncio
+async def test_s3_extractor_pages_and_reads_all_files(subject_with_populated_objects):
+    expected_results = [{"hello": i} for i in range(NUM_OBJECTS)]
+    results = [
+        result async for result in subject_with_populated_objects.extract_records()
+    ]
+    assert_that(results, has_length(NUM_OBJECTS))
+    assert_that(results, has_items(*expected_results))
+
+
+@pytest.mark.asyncio
+async def test_s3_extractor_pages_and_reads_all_files_with_object_format(
+    subject_with_populated_objects_and_no_extension,
+):
+    subject_with_populated_objects_and_no_extension.file_sources[0].object_format = (
+        ".json"
+    )
+    expected_results = [{"hello": i} for i in range(NUM_OBJECTS)]
+    results = [
+        result
+        async for result in subject_with_populated_objects_and_no_extension.extract_records()
+    ]
+    assert_that(results, has_length(NUM_OBJECTS))
+    assert_that(results, has_items(*expected_results))
+
+
+@pytest.mark.asyncio
+async def test_s3_extractor_pages_and_reads_all_gz_compressed_files(
+    subject_with_populated_and_gz_compressed_objects,
+):
+    expected_results = [{"hello": i} for i in range(NUM_OBJECTS)]
+    results = [
+        result
+        async for result in subject_with_populated_and_gz_compressed_objects.extract_records()
+    ]
+    assert_that(results, has_length(NUM_OBJECTS))
+    assert_that(results, has_items(*expected_results))
+
+
+@pytest.mark.asyncio
+async def test_s3_extractor_pages_and_reads_all_bz2_compressed_files(
+    subject_with_populated_and_bz2_compressed_objects,
+):
+    expected_results = [{"hello": i} for i in range(NUM_OBJECTS)]
+    results = [
+        result
+        async for result in subject_with_populated_and_bz2_compressed_objects.extract_records()
+    ]
+    assert_that(results, has_length(NUM_OBJECTS))
+    assert_that(results, has_items(*expected_results))
+
+
+@pytest.mark.asyncio
+async def test_s3_extractor_pages_and_reads_double_compressed_files(
+    subject_with_populated_double_compressed_objects,
+):
+    expected_results = [{"hello": i} for i in range(NUM_OBJECTS)]
+    results = [
+        result
+        async for result in subject_with_populated_double_compressed_objects.extract_records()
+    ]
+    assert_that(results, has_length(NUM_OBJECTS))
+    assert_that(results, has_items(*expected_results))
+
+
+@pytest.mark.asyncio
+async def test_s3_extractor_archives_objects(subject_with_archiving_enabled, s3_client):
+    [result async for result in subject_with_archiving_enabled.extract_records()]
+    assert_that(
+        s3_client.list_objects(Bucket=BUCKET_NAME, Prefix="archive", MaxKeys=1000)[
+            "Contents"
+        ],
+        has_length(NUM_OBJECTS),
+    )
+
+
+@pytest.mark.asyncio
+async def test_s3_extractor_does_not_archive_objects_when_archive_dir_is_none(
+    subject_with_populated_objects, s3_client
+):
+    [result async for result in subject_with_populated_objects.extract_records()]
+    assert_that(
+        s3_client.list_objects(Bucket=BUCKET_NAME, Prefix="archive", MaxKeys=1000),
+        not_(has_key("Contents")),
+    )
